feat: Cria retorno de erro personalizado para DataIntegrityViolationException com retorno de CONFLICT em ControllerExceptionHandler

+++ src/main/java/com/devsuperior/dsecommerce/services/ProductService.java
@@ -59,6 +59,10 @@ public class ProductService {
 
     @Transactional
     public void delete(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("ID " + id + " not found");
+        }
+
         productRepository.deleteById(id);
     }
